+++ src/GameCompleteScene.h
@@ -18,7 +18,7 @@ class GameCompleteScene : public Scene {
 public:
 	GameCompleteScene(ResourceManager* resourceManager);
 	void handleEvent(const SDL_Event& event) override;
-	void update(float deltaTime) override;
+	void update(Uint64 deltaTime) override;
 	void render(SDL_Renderer* renderer) override;
 private:
 	ResourceManager* m_resourceManager;
