+++ src/MenuScene.h
@@ -22,7 +22,7 @@ class MenuScene : public Scene {
 public:
 	MenuScene(ResourceManager* resourceManager);
 	void handleEvent(const SDL_Event& event) override;
-	void update(float deltaTime) override;
+	void update(Uint64 deltaTime) override;
 	void render(SDL_Renderer* renderer) override;
 private:
 	ResourceManager* m_resourceManager;
