+++ src/Scene.h
@@ -16,7 +16,7 @@ class Scene {
 public:
 	Scene() = default;
 	virtual void handleEvent(const SDL_Event& event) = 0;
-	virtual void update(float deltaTime) = 0;
+	virtual void update(Uint64 deltaTime) = 0;
 	virtual void render(SDL_Renderer* renderer) = 0;
 	virtual SceneResult getResult() const;
 	virtual ~Scene() = default;
